Корректровка ошибок работы с сетью

+++ app/src/main/java/com/example/recruiterhunter/presentation/seachVacancyVm/intents/SearchScreenSideEffects.kt
@@ -1,7 +1,7 @@
 package com.example.recruiterhunter.presentation.seachVacancyVm.intents
 
 sealed class SearchScreenSideEffects{
-    data class OpenDetails(val vacancyId: Int): SearchScreenSideEffects()
+    data class OpenDetails(val vacancyId: Long): SearchScreenSideEffects()
     data object OpenFilters: SearchScreenSideEffects()
     data object DownloadError: SearchScreenSideEffects()
 }
